// client/src/app/shared/models/products.ts
export interface IProduct {
    id: number;
    name: string;
    description: string;
    price: string;
    pictureUrl: string;
    productTypeName: string;
    productTypeId: number;
    productBrandName: string;
    productBrandId: number;
    totalRows: number;
    pageNum: number;
    rows: number;
}